updated structure for ghosts
