+++ PacmanProject/map.h
@@ -1,20 +1,5 @@
 #pragma once
-//Solution to course project # <5>
-//Introduction to programming course
-//Faculty of Mathematics and Informatics od Sofia University
-//Winter semester 2024/2025
-//
-//@author <Boris Dimitrov Tsvetkov>
-//idNumber <6MI0600504
-//@compiler Microsoft Visual Studio compiler
-//header file with helper variables
 
-const char wallSymbol = '#';
-const char foodSymbol = '@';
-const char pointSymbol = '.';
-const char emptySymbol = ' ';
-char** grid = nullptr;
-int widthGrid = 0, heightGrid = 0;
 
 /*void setConsoleCursorPosition(int x, int y);
 void updateGridCell(int x, int y, char symbol);
