+++ PacmanProject/x64/Debug/PacmanProject.tlog/Cl.items.tlog
@@ -1,1 +1,10 @@
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\BlueGhost.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\BlueGhost.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\Console.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\Console.obj
 C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\Game.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\Game.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\Ghost.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\Ghost.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\GreenGhost.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\GreenGhost.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\Main.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\Main.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\Pacman.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\Pacman.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\PinkGhost.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\PinkGhost.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\RedGhost.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\RedGhost.obj
+C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\Settings.cpp;C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\Settings.obj
