+++ PacmanProject/x64/Debug/PacmanProject.tlog/link.secondary.1.tlog
@@ -1,2 +1,2 @@
-^C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\GAME.OBJ
+^C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\BLUEGHOST.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\CONSOLE.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\GAME.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\GHOST.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\GREENGHOST.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\MAIN.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\PACMAN.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\PINKGHOST.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\REDGHOST.OBJ|C:\USERS\BOBI_\SOURCE\REPOS\PACMANPROJECT\PACMANPROJECT\X64\DEBUG\SETTINGS.OBJ
 C:\Users\bobi_\source\repos\PacmanProject\PacmanProject\x64\Debug\PacmanProject.ilk
